creating repositories and their implementations

+++ domain/src/main/java/com/github/slznvk/domain/dto/Coords.kt
@@ -1,6 +1,6 @@
 package com.github.slznvk.domain.dto
 
 data class Coords(
-    val lat: Int,
-    val long: Int
+    val lat: Double,
+    val long: Double
 )
